Fix incorrect length calculation in mdl_print_string_to_chan()
This fixes issue #52, where the score display in the endgame is
incorrectly broken into two lines. Line breaks are inserted between
strings (not between words), but the length for most characters was
counted twice.

The intention was to increase the length to account for non-printable
characters being converted into printable, with a "^" prefix, but part
of the condition had been inverted.